Rename voice_assistant to assist_pipeline
