+++ homeassistant/components/assist_pipeline/const.py
@@ -1,0 +1,2 @@
+"""Constants for the Assist pipeline integration."""
+DOMAIN = "assist_pipeline"
