+++ homeassistant/components/assist_pipeline/manifest.json
@@ -1,9 +1,9 @@
 {
-  "domain": "voice_assistant",
-  "name": "Voice Assistant",
+  "domain": "assist_pipeline",
+  "name": "Assist pipeline",
   "codeowners": ["@balloob", "@synesthesiam"],
   "dependencies": ["conversation", "stt", "tts"],
-  "documentation": "https://www.home-assistant.io/integrations/voice_assistant",
+  "documentation": "https://www.home-assistant.io/integrations/assist_pipeline",
   "iot_class": "local_push",
   "quality_scale": "internal",
   "requirements": ["webrtcvad==2.0.10"]
