+++ homeassistant/components/assist_pipeline/pipeline.py
