+++ homeassistant/components/assist_pipeline/vad.py
