--- homeassistant/components/voice_assistant/const.py
@@ -1,2 +1,0 @@
-"""Constants for the Voice Assistant integration."""
-DOMAIN = "voice_assistant"
