+++ requirements_all.txt
@@ -2619,7 +2619,7 @@ waterfurnace==1.1.0
 # homeassistant.components.cisco_webex_teams
 webexteamssdk==1.1.1
 
-# homeassistant.components.voice_assistant
+# homeassistant.components.assist_pipeline
 webrtcvad==2.0.10
 
 # homeassistant.components.whirlpool
