+++ requirements_test_all.txt
@@ -1880,7 +1880,7 @@ wallbox==0.4.12
 # homeassistant.components.folder_watcher
 watchdog==2.3.1
 
-# homeassistant.components.voice_assistant
+# homeassistant.components.assist_pipeline
 webrtcvad==2.0.10
 
 # homeassistant.components.whirlpool
