+++ tests/components/assist_pipeline/conftest.py
