+++ tests/components/assist_pipeline/snapshots/test_init.ambr
