+++ tests/components/assist_pipeline/snapshots/test_websocket.ambr
