+++ tests/components/assist_pipeline/test_init.py
