+++ tests/components/assist_pipeline/test_pipeline.py
