+++ tests/components/assist_pipeline/test_vad.py
