+++ tests/components/assist_pipeline/test_websocket.py
